refactor: extract reply-strategy-types.ts to break circular imports
Break circular import cycles between reply-strategy.ts and its concrete
implementations (reply-strategy-card.ts, reply-strategy-markdown.ts) (#519).

The shared interfaces (DeliverPayload, ReplyOptions, ReplyStrategy,
ReplyStrategyContext) now live in a leaf module that nothing in the cycle
imports back from. reply-strategy.ts re-exports them for backward compat.
